+++ npins.nix
@@ -51,6 +51,10 @@ let
     version = cargoToml.package.version;
     cargoLock = {
       lockFile = src + "/Cargo.lock";
+
+      outputHashes = {
+        "nix-compat-0.1.0" = "sha256-U9pAde6R2yoP8ivnoNX/1rve+ALrDk8+4R2BKoGzg24=";
+      };
     };
 
     inherit src;