+++ src/pypi.rs
@@ -1,6 +1,6 @@
 //! Pin a PyPi package
 
-use crate::*;
+use crate::{nix::hash_to_sri, *};
 use anyhow::{Context, Result};
 use lenient_version::Version;
 use serde::{Deserialize, Serialize};
@@ -125,11 +125,15 @@ impl Updatable for Pin {
                 anyhow::format_err!("Unsupported package: must contain some \"source\" download",)
             })?;
 
-        let hash = latest_source.digests.remove("sha256").ok_or_else(|| {
-            anyhow::format_err!(
-                "JSON metadata is invalid: must contain a `sha256` entry within `digests`",
-            )
-        })?;
+        let hash = latest_source
+            .digests
+            .remove("sha256")
+            .ok_or_else(|| {
+                anyhow::format_err!(
+                    "JSON metadata is invalid: must contain a `sha256` entry within `digests`",
+                )
+            })
+            .and_then(|s| hash_to_sri(&s, "sha256"))?;
 
         Ok(GenericUrlHashes {
             hash,
@@ -190,7 +194,7 @@ mod test {
         assert_eq!(
             pin.fetch(&version).await?,
             GenericUrlHashes {
-                hash: "3953b158b7b690642d68cd6beb1d59f6e10526f2ee10a6fb4636a913cc95e718".into(),
+                hash: "sha256-OVOxWLe2kGQtaM1r6x1Z9uEFJvLuEKb7RjapE8yV5xg=".into(),
                 url: "https://files.pythonhosted.org/packages/d1/d5/0c270c22d61ff6b883d0f24956f13e904b131b5ac2829e0af1cda99d70b1/gaiatest-0.34.tar.gz".parse().unwrap(),
             }
         );
@@ -216,7 +220,7 @@ mod test {
         assert_eq!(
             pin.fetch(&version).await?,
             GenericUrlHashes {
-                hash: "39d09c6627255fcf39c938937995665b6377799c4fa141f6b481bcb5e6a688ac".into(),
+                hash: "sha256-OdCcZiclX885yTiTeZVmW2N3eZxPoUH2tIG8teamiKw=".into(),
                 url: "https://files.pythonhosted.org/packages/fd/75/6e72889c3b154a179040b94963a50901966ff30b68600271df374b2ded7a/streamlit-0.89.0.tar.gz".parse().unwrap(),
             }
         );